Update SalesAssociate.java
added a file name for the Associate to choose the name for the bike parts to be stored. This name will have to be stored in an arraylist somewhere to be retrieved when making the profile after closer. Also created a file name for the file to contain the transactions.